fix: migrate long to integer

+++ src/main/java/com/example/spring/model/AutoComplete/LegalForm.java
@@ -15,7 +15,6 @@ public class LegalForm {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private Integer id;
     private String name;
 }
